+++ Assets/Scripts/LevelData.cs
@@ -8,11 +8,9 @@ public class LevelData : MonoBehaviour
 {
     [SerializeField] private float maxScore=200;
     [SerializeField] private float startingScore=100;
-    [SerializeField] private float levelTime=300;
 
     private void Awake()
     {
         EventManager.currentManager.AddEvent(new SetScoreStats(maxScore,startingScore));
-        EventManager.currentManager.AddEvent(new SetTime(levelTime));
     }
 }
